requirements.txt
Django==2.2
django-allauth==0.39.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
mysqlclient==1.4.2.post1
